Remove kustomize CLI and use module again
